Add Role Application Service

+++ src/Identity Context/Ledger.Identity.Application/AppServices/RoleAppServices/RoleApplicationService.cs
@@ -1,6 +1,7 @@
 ﻿using Ledger.CrossCutting.ServiceBus.Abstractions;
 using Ledger.Identity.Domain.Aggregates.RoleAggregate;
 using Ledger.Identity.Domain.Commands.RoleCommands;
+using Ledger.Identity.Domain.Models.Services.UserServices;
 using Ledger.Identity.Domain.Services.RoleServices;
 using Ledger.Shared.IntegrationEvents.Events.RoleEvents;
 using Ledger.Shared.Notifications;
@@ -13,10 +14,12 @@ namespace Ledger.Identity.Application.AppServices.RoleAppServices
     public class RoleApplicationService : BaseApplicationService, IRoleApplicationService
     {
         private readonly LedgerRoleManager _roleManager;
+        private readonly LedgerUserManager _userManager;
 
-        public RoleApplicationService(LedgerRoleManager roleManager, IDomainNotificationHandler domainNotificationHandler, IDomainServiceBus domainServiceBus, IIntegrationServiceBus integrationBus) : base(domainNotificationHandler, domainServiceBus, integrationBus)
+        public RoleApplicationService(LedgerRoleManager roleManager, LedgerUserManager userManager, IDomainNotificationHandler domainNotificationHandler, IDomainServiceBus domainServiceBus, IIntegrationServiceBus integrationBus) : base(domainNotificationHandler, domainServiceBus, integrationBus)
         {
             _roleManager = roleManager;
+            _userManager = userManager;
         }
 
         public IQueryable<LedgerIdentityRole> GetAllRoles()
@@ -30,7 +33,7 @@ namespace Ledger.Identity.Application.AppServices.RoleAppServices
             return role;
         }
 
-        public async Task AddRole(RegisterRoleCommand command)
+        public async Task Register(RegisterRoleCommand command)
         {
             command.Validate();
 
@@ -47,7 +50,7 @@ namespace Ledger.Identity.Application.AppServices.RoleAppServices
             return;
         }
 
-        public async Task RemoveRole(RemoveRoleCommand command)
+        public async Task Remove(RemoveRoleCommand command)
         {
             command.Validate();
 
@@ -58,6 +61,13 @@ namespace Ledger.Identity.Application.AppServices.RoleAppServices
 
             if (NotifyNullRole(role))
                 return;
+
+            bool canDelete = await CanDeleteRole(role);
+            if (!canDelete)
+            {
+                AddNotification("Falha ao remover", "Não é possível remover essa role enquanto ouver usuários nela.");
+                return;
+            }
 
             IdentityResult result = await _roleManager.DeleteAsync(role);
             if (result.Succeeded)
@@ -76,6 +86,11 @@ namespace Ledger.Identity.Application.AppServices.RoleAppServices
             }
 
             return false;
+        }
+
+        private async Task<bool> CanDeleteRole(LedgerIdentityRole role)
+        {
+            return !(await _userManager.GetUsersInRoleAsync(role.Name)).Any();
         }
     }
 }
